GriseldaStudioAPI/requirements.txt
asgiref==3.8.1
azure-core==1.30.2
azure-storage-blob==12.20.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==42.0.8
Django==5.0.6
django-cors-headers==4.3.1
django-filter==24.2
django-storages==1.14.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-nested-routers==0.94.1
gunicorn==22.0.0
idna==3.7
iniconfig==2.0.0
isodate==0.6.1
packaging==24.0
pillow==10.3.0
pluggy==1.5.0
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
pytest==8.2.0
pytest-django==4.8.0
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
whitenoise==6.6.0
